feat(acc): bump recorder to v3 + capture full SPageFileGraphic
ACC shared memory graphics struct was truncated at 1320 bytes (legacy SDK
constant), dropping everything past rainTyres — including sessionIndex,
deltaLapTime, isValidLap, missingMandatoryPits, clock. Current Kunos SDK
(v1.8.12) exposes 1588 bytes.

- structs.ts: SIZE 1320 → 1588; add MIN_SIZE=1320 so legacy bins keep parsing;
  map tail offsets (sessionIndex, usedFuel, iDeltaLapTime, iEstimatedLapTime,
  isDeltaPositive, iSplit, isValidLap@1408, fuelEstimatedLaps,
  missingMandatoryPits, clock).
- recorder.ts: bump format version 2 → 3. Header layout unchanged. V2 bins
  still readable (frames self-describe their size).
- frame-reader.ts: accept versions {2, 3}; warn + skip unknown.
- parser.ts: gate on MIN_SIZE; read isValidLap with length guard (null on V2,
  boolean on V3). No downstream consumer yet.
- shared/types.ts: AccExtendedData.isValidLap: boolean | null.
- acc-routes.ts, replay.ts: in-flight work included in this commit.
